refactor: centralize token parsing
